make shared things utils
